fix build

+++ src/server/game/Globals/ObjectMgr.h
@@ -738,7 +738,7 @@ class ObjectMgr
         CreatureModelInfo const* GetCreatureModelRandomGender(uint32* displayID);
         static uint32 ChooseDisplayId(CreatureTemplate const* cinfo, CreatureData const* data = NULL);
         static void ChooseCreatureFlags(CreatureTemplate const* cinfo, uint64& npcflag, uint32& unit_flags, uint32& dynamicflags, CreatureData const* data = NULL);
-        EquipmentInfo const* GetEquipmentInfo(uint32 entry, int8& id);
+        EquipmentInfo const* GetEquipmentInfo(uint32 entry, int32& id);
         CreatureAddon const* GetCreatureAddon(ObjectGuid::LowType lowguid);
         GameObjectAddon const* GetGameObjectAddon(ObjectGuid::LowType lowguid);
         CreatureAddon const* GetCreatureTemplateAddon(uint32 entry);
